requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo>=4.3.0
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn
python-multipart


## -e .
